+++ src/components/organisms/KineticTypography/gl/frag.glsl
