+++ src/components/organisms/KineticTypography/useText.ts
@@ -1,10 +1,4 @@
-import {
-  MutableRefObject,
-  useCallback,
-  useEffect,
-  useMemo,
-  useState,
-} from "react";
+import { useCallback, useMemo, useState } from "react";
 import initCanvas from "@/lib/utils/canvas/initCanvas";
 
 const fontSize = 70;
